Setting up the base primitives
Points are just that, points the laser can draw. Frames are collections of points. Screens are a collection of frames with the added twist that frames can be a subset of the drawable area by the laser. Eventually this will allow me to use frames to represent individual characters or shapes and move them without directly altering the state of the screen object.